aws/requirements.txt
aws-cdk.core==1.69.0
aws-cdk.aws_s3==1.69.0
aws-cdk.aws_s3_deployment==1.69.0
aws-cdk.aws_iam==1.69.0
aws-cdk.aws_cloudfront==1.69.0
aws-cdk.aws_certificatemanager==1.69.0
aws-cdk.aws_route53==1.69.0
aws-cdk.aws_route53_targets==1.69.0
aws-cdk.aws_dynamodb==1.69.0
aws-cdk.aws_lambda==1.69.0
aws-cdk.aws_appsync==1.69.0
aws_cdk.aws_batch==1.69.0
aws_cdk.aws_ecr==1.69.0
aws_cdk.aws_ecs==1.69.0
aws_cdk.aws_ec2==1.69.0